+++ fj-doc-val-p7m/src/main/java/org/fugerit/java/doc/val/p7m/P7MValidator.java
@@ -1,12 +1,13 @@
 package org.fugerit.java.doc.val.p7m;
 
-import java.io.InputStream;
-
-import org.bouncycastle.cms.CMSSignedData;
+import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.doc.val.core.DocTypeValidationResult;
 import org.fugerit.java.doc.val.core.DocTypeValidator;
 import org.fugerit.java.doc.val.core.basic.AbstractDocTypeValidator;
 
+import java.io.InputStream;
+
+@Slf4j
 public class P7MValidator extends AbstractDocTypeValidator {
 	
 	public static final String EXTENSION = "P7M";
@@ -18,10 +19,10 @@ public class P7MValidator extends AbstractDocTypeValidator {
 	public P7MValidator() {
 		super( MIME_TYPE, EXTENSION );
 	}
-	
+
 	@Override
 	public DocTypeValidationResult validate(InputStream is) {
-		return this.validationHelper( () -> new CMSSignedData( is ) );
+		return this.validationHelper(() -> P7MUtils.extractContent( is ) );
 	}
-	
+
 }